Refector Generator class
Split it in 2 classes:
BuildGenerator: previously Generator
ProjectGenerator: focuses on a single gradle project.

This is to prepare extracting actual build file generation
from it so that we can handle groovy vs kts.
