+++ generator/src/main/kotlin/com/android/gradle/replicator/generator/ProjectGenerator.kt
@@ -17,89 +17,20 @@
 
 package com.android.gradle.replicator.generator
 
-import com.android.gradle.replicator.model.*
-import com.android.gradle.replicator.model.internal.DefaultDependenciesInfo
+import com.android.gradle.replicator.model.DependenciesInfo
+import com.android.gradle.replicator.model.ModuleInfo
+import com.android.gradle.replicator.model.PluginType
+import com.android.gradle.replicator.model.ProjectInfo
 import java.io.File
 
-class Generator(private val params: Params) {
-    interface Params {
-        val jsonFile: File
-        val destination: File
-        val libraryFilter: File?
-        val libraryAdditions: File?
-    }
+class ProjectGenerator(
+    private val destinationFolder: File,
+    private val libraryFilter: Map<String, String>,
+    private val libraryAdditions: Map<String, List<DependenciesInfo>>
+) {
 
-    private val libraryFilter = generateLibraryFilter()
-    private val libraryAdditions = generateLibraryAdditions()
-
-    fun generate() {
-        val project = Serializer.instance().deserializeProject(params.jsonFile)
-
-        println("Project: ':'")
-        generateRootModule(project)
-
-        val count = project.subModules.size
-        val digitCount = count.toString().length
-        val formatter = "%0${digitCount}d"
-
-        var index = 1
-        for (module in project.subModules) {
-            println("Project(${formatter.format(index)}): '${module.path}'")
-            generateModule(getFolder(module.path), module)
-            index++
-        }
-
-        generateSettingsFile(project)
-        generateGradleProperties(project.gradleProperties)
-
-        println("Done.")
-    }
-
-    private data class RootPluginInfo(
-        val plugin: PluginType,
-        val applied: Boolean)
-
-    private fun generateLibraryFilter(): Map<String, String> {
-        return params.libraryFilter?.let { file: File ->
-            file.readLines().filter { it.isNotEmpty() }.map {
-                val split = it.split(" -> ")
-                if (split.size == 1) {
-                    split[0] to ""
-                } else {
-                    split[0] to split[1]
-                }
-            }.toMap()
-        } ?: mapOf()
-    }
-
-    private fun generateLibraryAdditions(): Map<String, List<DependenciesInfo>> {
-        return params.libraryAdditions?.let { file: File ->
-            val entries = file.readLines().filter { it.isNotEmpty() }.map {
-                val split = it.split(" ")
-                if (split.size != 3) {
-                    throw RuntimeException("badly formatter library addition line: '$it'")
-                }
-
-                split[0] to DefaultDependenciesInfo(
-                    scope = split[1],
-                    dependency = split[2],
-                    type = DependencyType.EXTERNAL_LIBRARY
-                )
-            }
-
-            val result = mutableMapOf<String, List<DependenciesInfo>>()
-
-            entries.forEach {
-                val list = result.computeIfAbsent(it.first) { mutableListOf() } as MutableList<DependenciesInfo>
-                list.add(it.second)
-            }
-
-            result
-        } ?: mapOf()
-    }
-
-    private fun generateRootModule(project: ProjectInfo) {
-        val buildFile = File(params.destination, "build.gradle")
+    internal fun generateRootModule(project: ProjectInfo) {
+        val buildFile = File(destinationFolder, "build.gradle")
 
         val rootPlugins = project.rootModule.plugins
 
@@ -119,26 +50,26 @@ class Generator(private val params: Params) {
         val pluginIds = rootPluginInfos.asSequence()
             .filter { it.plugin.useNewDsl }
             .map {
-            val applyStr = if (it.applied) {
-                ""
-            } else {
-                " apply false"
-            }
+                val applyStr = if (it.applied) {
+                    ""
+                } else {
+                    " apply false"
+                }
 
-            when (it.plugin) {
-                PluginType.JAVA, PluginType.JAVA_LIBRARY, PluginType.APPLICATION -> {
-                    if (it.applied) {
-                        """    id("${it.plugin.id}")"""
-                    } else {
-                        ""
+                when (it.plugin) {
+                    PluginType.JAVA, PluginType.JAVA_LIBRARY, PluginType.APPLICATION -> {
+                        if (it.applied) {
+                            """    id("${it.plugin.id}")"""
+                        } else {
+                            ""
+                        }
                     }
+                    PluginType.KOTLIN_ANDROID, PluginType.KOTLIN_JVM, PluginType.KAPT -> {
+                        """    id("${it.plugin.id}") version "${project.kotlinVersion}" $applyStr """
+                    }
+                    else -> throw RuntimeException("Unexpected plugin in root plugin infos.")
                 }
-                PluginType.KOTLIN_ANDROID, PluginType.KOTLIN_JVM, PluginType.KAPT -> {
-                    """    id("${it.plugin.id}") version "${project.kotlinVersion}" $applyStr """
-                }
-                else -> throw RuntimeException("Unexpected plugin in root plugin infos.")
-            }
-        }.joinToString(separator = "\n")
+            }.joinToString(separator = "\n")
 
         // now prepare the classpaths for plugins *not* using the new DSL
         val classpaths = allPlugins.asSequence()
@@ -183,10 +114,10 @@ allprojects {
 """)
 
         // now the generic module info stuff
-        generateModuleInfo(params.destination, project.rootModule)
+        generateModuleInfo(destinationFolder, project.rootModule)
     }
 
-    private fun generateModule(
+    internal fun generateModule(
         folder: File,
         module: ModuleInfo
     ) {
@@ -225,6 +156,7 @@ allprojects {
         module.generateDependencies(buildFile)
     }
 
+
     private fun ModuleInfo.generateDependencies(
         buildFile: File
     ) {
@@ -261,23 +193,9 @@ allprojects {
         buildFile.appendText("}\n")
     }
 
-    private fun generateSettingsFile(project: ProjectInfo) {
-        println("Generate settings.gradle")
-        val settingsGradle = File(params.destination, "settings.gradle")
-        settingsGradle.writeText(project.subModules.map { "include(\"${it.path}\")" }.sorted().joinToString("\n"))
-    }
 
-    private fun generateGradleProperties(properties: List<String>) {
-        if (properties.isEmpty()) return
+    private data class RootPluginInfo(
+        val plugin: PluginType,
+        val applied: Boolean)
 
-        println("Generate gradle.properties")
-        val settingsGradle = File(params.destination, "gradle.properties")
-        settingsGradle.writeText(properties.sorted().joinToString("\n"))
-    }
-
-    private fun getFolder(path: String): File {
-        return params.destination.join(path.split(":")).also {
-            it.createDirWithParents()
-        }
-    }
 }
